rename & refactor envHelper as generic class;
