--- Runtime/DTOs/EnvCustomTicketsRequestDTO.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 943594dd800edb74abc4c2505e9f29d0
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
