--- Runtime/DTOs/EnvEqualityDTO.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 4150fb8f592578b4995cb304eea49d40
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
