--- Runtime/DTOs/EnvIntersectionDTO.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: cebdcd056cc0a434e935ccfb9d2d3483
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
